able to save & recreate state from storage
- Added a wrapping struct for web_sys::Url to enable deserialize
- Misc updates related to above

TODO: Listen for changes in the app to persist state. The current usage
of signals is with mutable inside of the sections. This might be hard
react to on the application layer. The solution might be to use events
used inside of sections, to propagate these events up and then listen
for this event, and finally save the state.